chore: move all components under lib/components
