+++ src/routes/+page.ts
@@ -1,4 +1,3 @@
-import type { IArticleService, IHomePageService, IMenuService } from '$lib/types/b3';
 import type { HomePageData } from '$lib/types/b3.pagedata';
 import { HomePageService } from '$lib/services/contentful/HomePageService';
 import client from '$lib/infrastructure/contentful/graphQLClient';
@@ -6,9 +5,9 @@ import { ArticleService } from '$lib/services/contentful/ArticleService';
 import { MenuService } from '$lib/services/contentful/MenuService';
 
 export async function load(): Promise<HomePageData> {
-	const homePageService: IHomePageService = new HomePageService(client);
-	const articleService: IArticleService = new ArticleService(client);
-	const menuService: IMenuService = new MenuService(client);
+	const homePageService = new HomePageService(client);
+	const articleService = new ArticleService(client);
+	const menuService = new MenuService(client);
 
 	const [topBanner, articles, menuItems] = await Promise.all([
 		homePageService.getTopBanner(),
